rename create_training_data.py to training_data.py
