+++ bin/create_training_data.py
@@ -3,8 +3,8 @@
 import argparse
 
 from src.download_naips import NAIPDownloader
-from src.create_training_data import (random_training_data, equalize_data, split_train_test,
-                                      format_as_onehot_arrays, dump_data_to_disk)
+from src.training_data import (random_training_data, equalize_data, split_train_test,
+                               format_as_onehot_arrays, dump_data_to_disk)
 
 
 def create_parser():
